a method of localizing a database
when using the sdk at high frequencies with large queries everything slows down exponentially, and eventually locks you out for a time. this short optional script would create and write to the database all cards and the data needed for whatever reason (in my case it will be read by a diagnostic for deck building reasons) the script is malleable, easy to follow, and im sure more than a few users have probably written it themselves. however, do to the restrictive nature of the server, and the fact i already have it written i figured id post it up. definitely worth the time.